test(integration-create-get-invoice): add invalid create invoice request tests
invalid create and get invoice requests added.
