+++ api/package.json
@@ -23,7 +23,6 @@
   },
   "jest": {
     "testPathIgnorePatterns": [
-      "/src/__tests__/integration/",
       "/node_modules/"
     ]
   }
